requirements.txt
torch
torchaudio
torchvision
transformers
accelerate
bitsandbytes
sentencepiece
huggingface_hub
datasets

speechbrain
librosa
soundfile
edge-tts

opencv-python
numpy
Pillow

gradio
python-dotenv
azure-cognitiveservices-speech
